Moving Tk stuff to interface.py
